通報チャンネルをリアクションで作成できるように修正

+++ app/suggestion-box.js
@@ -1,50 +1,60 @@
 const BOT_ROLE_NAME = "bot";
 const common = require("./common.js");
+const suggestionChannelParentId = process.env.CATEGORY_PARENT_ID_SUGGESTION_BOX;
+const infomationChannelId = process.env.CHANNEL_ID_INFOMATION;
+const Discord = require("discord.js");
 
-module.exports = async function suggestionBox(msg) {
-    const guild = msg.guild;
-    const suggestionChannel = await guild.channels.cache.find(
-        (channel) => channel.name === "ご意見や通報などはこちら"
-    );
-
-    if (msg.channel.name === suggestionChannel.name) {
-        if (!msg.member.hasPermission("ADMINISTRATOR")) {
-            const msgContent = msg.content;
-            const newChannel = await txChCreate(msg);
-            newChannel.send("@everyone " + `<@${msg.author.id}>` +"さんからご意見箱に投稿があったでし");
-            newChannel.send("```" + msg.content + "```");
+module.exports = {
+    init: async function suggestionBox(msg, user) {
+        if (msg.member.hasPermission("ADMINISTRATOR")
+            && suggestionChannelParentId == msg.channel.parent.id
+            && msg.content === "!init") {
+            let messgeEmbed = new Discord.MessageEmbed()
+                .setTitle('ご意見や通報、お問い合わせについて')
+                .setColor(0x008080)
+                .setDescription('イカ部に関するご意見や、メンバーからの嫌がらせなどを通報するにはこのメッセージに📭リアクションをしてください。\n' +
+                    'あなたと管理者だけしか閲覧できないプライベートチャンネルが作成されます。\n' +
+                    '頂いたご意見、通報内容の全てに対応できる訳ではございませんので、ご了承くださいませ。' +
+                    `ご意見の一部は <#${infomationChannelId}> にて回答させて頂く場合がございます。`);
+            msg.channel.send(messgeEmbed).then((sentMessage) =>
+                sentMessage.react('📭'));
             msg.delete();
-            return true;
         }
-    } else if (msg.member.hasPermission("ADMINISTRATOR")
-        && suggestionChannel.parent == msg.channel.parent
-        && msg.content === "!close") {
-        const sendChannel = await guild.channels.cache.find(
-            (channel) => channel.name === "ご意見箱"
-        );
-        msg.delete();
-
-        let messages = await msg.channel.messages.fetch({ limit: 100 }).catch(console.error);
-        messages.sort(function (a, b) {
-            if (a.createdTimestamp < b.createdTimestamp) return -1;
-            if (a.createdTimestamp > b.createdTimestamp) return 1;
-            return 0;
-        });
-
-        messages.map(
-            m => sendChannel.send(common.composeEmbed(m))
-        );
-        txChHide(msg);
-        return true;
+    },
+    create: async function suggestionBox(msg, user) {
+        const newChannel = await txChCreate(msg, user);
+        newChannel.send("＠everyone " + `<@${user.id}>` + "さん専用のチャンネルでし\n問い合わせ内容を入力してほしいでし\n`(管理者のみ)!close`で問い合わせを終了してアーカイブできるでし");
+        return;
+    },
+    archive: async function suggestionBoxArchive(msg) {
+        if (msg.member.hasPermission("ADMINISTRATOR")
+            && suggestionChannelParentId == msg.channel.parent.id
+            && msg.content === "!close") {
+            msg.delete();
+            let messages = await msg.channel.messages.fetch({ limit: 100 }).catch(console.error);
+            messages.sort(function (a, b) {
+                if (a.createdTimestamp < b.createdTimestamp) return -1;
+                if (a.createdTimestamp > b.createdTimestamp) return 1;
+                return 0;
+            });
+            const guild = msg.guild;
+            const sendChannel = await guild.channels.cache.find(
+                (channel) => channel.id === process.env.CHANNEL_ID_SUGGESTION_BOX_ARCHIVE
+            );
+            messages.map(
+                m => sendChannel.send(common.composeEmbed(m))
+            );
+            txChHide(msg);
+            return;
+        }
     }
-    return false;
 }
 
-async function txChCreate(msg) {
+async function txChCreate(msg, user) {
     try {
         const guild = msg.channel.guild;
         const msgChannel = msg.channel;
-        let chName = await msg.createdTimestamp + "-" + msg.author.username;
+        let chName = await msg.createdTimestamp + "-" + user.username;
         let botRole = await guild.roles.cache.find(val => val.name === BOT_ROLE_NAME);
         let result = await guild.channels.create(chName, {
             parent: msgChannel.parent,
